new emitter config 'report_all_errors' & implement this for perl
the default emitted validator code returns a hash {success=>0|1, errors=>[...],
warnings=>[...]}. for very simple schemas (like 'int', 'str*') this is an
overkill (the creation of hashes and arrays of results), so we add an option to
generate leaner code. if 'report_all_errors' is set to false, the emitted code
will just return a single error string (or undef if no error is found).